requirements.txt
# Core
fastapi
uvicorn
pydantic
python-dotenv
python-multipart

# Database
psycopg2-binary
sqlalchemy
alembic
redis

# Testing
pytest

# Integrations
openai
requests
twilio
pyngrok
stripe